feat(auto-give): 增加首次加入自动给予物品功能并优化冷却机制
- 新增首次加入自动给予物品配置选项
- 实现首次加入物品给予逻辑
- 优化冷却时间计算方式，基于玩家游戏时间
- 调整冷却时间配置，支持小数
- 修复 Spigot API 版本不兼容问题
- 优化颜色转换逻辑，支持 1.13 以下版本

+++ src/main/java/cn/ningmo/bellcommand/item/CommandItem.java
@@ -16,7 +16,7 @@ public class CommandItem {
     private final String name;
     private final List<String> lore;
     private final String permission;
-    private final int cooldown;
+    private final double cooldown;
     private final Map<String, List<CommandEntry>> commands;
 
     public static class CommandEntry {
@@ -51,7 +51,7 @@ public class CommandItem {
         this.lore = ColorUtils.translateColors(rawLore);
         
         this.permission = config.getString("permission", "");
-        this.cooldown = config.getInt("cooldown", 0);
+        this.cooldown = config.getDouble("cooldown", 0.0);
         this.commands = new HashMap<>();
         
         // 加载命令
@@ -93,7 +93,7 @@ public class CommandItem {
         return permission;
     }
 
-    public int getCooldown() {
+    public double getCooldown() {
         return cooldown;
     }
